Fixed: find, find_if, find_if_not would always return the first item in the list if a match was found and iterators were used for range

+++ include/gaia/ecs/query_info.h
@@ -53,8 +53,8 @@ namespace gaia {
 							return false;
 
 						// Read-write mask must match
-						const uint8_t maskRW = (uint32_t)data.readWriteMask & (1U << (uint32_t)idx);
-						const uint8_t maskXX = (uint32_t)isReadWrite << idx;
+						const uint32_t maskRW = (uint32_t)data.readWriteMask & (1U << (uint32_t)idx);
+						const uint32_t maskXX = (uint32_t)isReadWrite << idx;
 						return maskRW == maskXX;
 					}
 				}
